+++ src/ConnectionManager.h
@@ -30,13 +30,18 @@ class ConnectionManager final {
 
   bool isListening(uint16_t port);
 
-  // Finds a connection whose local port is the specified port.
+  // Stop listening on the specified port. This returns true if the listener was
+  // found and successfully stopped. This returns false if the listener was not
+  // found or was found and not successfully stopped.
+  bool stopListening(uint16_t port);
+
+  // Find a connection whose local port is the specified port.
   std::shared_ptr<ConnectionHolder> findConnected(uint16_t port);
 
-  // Finds a connection on the given port that has data available.
+  // Find a connection on the given port that has data available.
   std::shared_ptr<ConnectionHolder> findAvailable(uint16_t port);
 
-  // Removes the given connection and returns whether the connection existed in
+  // Remove the given connection and returns whether the connection existed in
   // the list and was removed.
   bool remove(const std::shared_ptr<ConnectionHolder> &holder);