+++ CHANGELOG.md
@@ -4,6 +4,51 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0-snapshot]
+
+### Added
+* Added a new "survey of how connections work" section to the README.
+* Added low-level link receive error stats collection.
+* Now calling `yield()` in `EthernetUDP::parsePacket()` when it returns zero so
+  that calls in a loop will move the stack forward.
+
+### Changed
+* Changed `EthernetClass::loop()` to be `static`.
+* Changed all the internal "`yield()` to move the stack along" calls to
+  `EthernetClass::loop()`. Note that the calls within `while` loops in the
+  external API functions were not changed.
+
+## [0.5.0]
+
+### Added
+* Added a "Code style" section to the README.
+* Added link-status and address-changed callbacks.
+* New `EthernetServer::end()` function to stop listening.
+* Disabling the PLL before disabling the clock in `enet_deinit()`. This still
+  doesn't solve the freeze problem when this function is called.
+* New `EthernetClass::linkSpeed()` function, used as `Ethernet.linkSpeed()`.
+  It returns the link speed in Mbps.
+
+### Changed
+* Changed all the delays to yields because `delay()` just loops and calls
+  `yield()` anyway.
+* No longer looping when checking the TCP send buffer in the client write
+  functions. Instead, just using an `if`.
+* The client now takes more opportunities to set the internal connection pointer
+  to NULL, meaning the caller doesn't necessarily need to guarantee they call
+  `stop()` if other I/O functions are being used to check state or to write,
+  for example.
+* Moved adding the netif callback to `enet_init()` so that the callback still
+  gets an address-changed notification when setting a static address.
+
+### Fixed
+* Fixed client functions to also check for connected status.
+* Fixed client `read()` and `peek()` to return -1 on no connection.
+* Added potential flushing in `EthernetClient::availableForWrite()`. This keeps
+  things moving along if it always would return zero.
+* A listening server is now correctly added to the internal listening list. This
+  fixes `EthernetServer::operator bool()`.
+
 ## [0.4.0]
 
 ### Added
@@ -46,10 +91,12 @@ and this project adheres to
   to avoid having to call `yield()` themselves.
 
 ### Changed
-* Removed all the atomic fences.
 * Brought `Print::write` functions into scope for Client, Server, and UDP by
   using a `using Print::write` directive.
 * New centralized connection management.
+
+### Removed
+* Removed all the atomic fences.
 
 ### Fixed
 * Fixed `EthernetClass::begin()` return value; it was the opposite.